requirements.txt
easydict==1.6
Jinja2==2.8
MarkupSafe==0.23
py2-ipaddress==3.4.1
PyYAML==3.11
wheel==0.29.0
